Updated to delete all existing project resource mapping

+++ src/main/java/com/capacityplanning/resourceutilization/repository/ProjectResourceMappingRepository.java
@@ -90,5 +90,5 @@ public interface ProjectResourceMappingRepository extends JpaRepository<ProjectR
     List<ProjectResourceMappingEntity> findByStatus(@Param("status") String status);
 
     @Query("DELETE FROM ProjectResourceMappingEntity prm WHERE prm.projectId = :projectId AND prm.resourceId = :resourceId")
-    void deleteByProjectIdAndResourceId(@Param("projectId") Long projectId, @Param("resourceId") Long resourceId);
+    void deleteByProjectIdAndResourceId(@Param("projectId") Integer projectId, @Param("resourceId") Integer resourceId);
 }
